Move metronome into own class and rename to avoid conflict with mod sounds
